+++ src/components/v5/common/ActionSidebar/partials/AmountField/AmountField.tsx
@@ -1,5 +1,3 @@
-import { Id } from '@colony/colony-js';
-import { WarningCircle } from '@phosphor-icons/react';
 import {
   DefaultTimeDelimiter,
   formatNumeral,
@@ -14,26 +12,20 @@ import React, {
   useEffect,
   useRef,
 } from 'react';
-import { useController } from 'react-hook-form';
+import { useController, useFormContext } from 'react-hook-form';
 
 import { useAdditionalFormOptionsContext } from '~context/AdditionalFormOptionsContext/AdditionalFormOptionsContext.ts';
 import { useColonyContext } from '~context/ColonyContext/ColonyContext.ts';
 import useRelativePortalElement from '~hooks/useRelativePortalElement.ts';
 import useToggle from '~hooks/useToggle/index.ts';
-import Tooltip from '~shared/Extensions/Tooltip/Tooltip.tsx';
-import Numeral from '~shared/Numeral/Numeral.tsx';
 import { formatText } from '~utils/intl.ts';
-import { multiLineTextEllipsis } from '~utils/strings.ts';
-import {
-  getBalanceForTokenAndDomain,
-  getTokenDecimalsWithFallback,
-} from '~utils/tokens.ts';
-import HoverWidthWrapper from '~v5/shared/HoverWidthWrapper/HoverWidthWrapper.tsx';
 import MenuContainer from '~v5/shared/MenuContainer/index.ts';
 import Portal from '~v5/shared/Portal/index.ts';
-import { TokenAvatar } from '~v5/shared/TokenAvatar/TokenAvatar.tsx';
 
 import { useAmountField } from './hooks.ts';
+import { TokenItem } from './partials/TokenItem.tsx';
+import { TokenList } from './partials/TokenList.tsx';
+import { TokenNotFound } from './partials/TokenNotFound.tsx';
 import { type AmountFieldProps } from './types.ts';
 
 const displayName = 'v5.common.ActionsContent.partials.AmountField';
@@ -55,6 +47,7 @@ const AmountField: FC<AmountFieldProps> = ({
   const placeholderValue =
     placeholder || formatText({ id: 'actionSidebar.enterAmount' });
 
+  const { trigger } = useFormContext();
   const {
     field,
     fieldState: { error },
@@ -105,6 +98,14 @@ const AmountField: FC<AmountFieldProps> = ({
     setValue(formatNumeral(e.target.value, formattingOptions));
   };
 
+  const handleTokenSelect = (selectedTokenAddress: string) => {
+    tokenAddressController.onChange(selectedTokenAddress);
+    if (value) {
+      trigger(name);
+    }
+    toggleTokenSelect();
+  };
+
   useEffect(() => {
     if (!inputRef.current) {
       return undefined;
@@ -148,37 +149,9 @@ const AmountField: FC<AmountFieldProps> = ({
   );
 
   const selectedTokenContent = isTokenInColony ? (
-    <div className="flex items-center gap-1">
-      <TokenAvatar
-        size={18}
-        tokenName={activeToken.name}
-        tokenAddress={activeToken.tokenAddress}
-        tokenAvatarSrc={activeToken.avatar ?? undefined}
-      />
-      <span
-        className={clsx('text-md', {
-          'text-gray-300': isDisabled,
-          'text-negative-400': tokenAddressError,
-        })}
-      >
-        {multiLineTextEllipsis(activeToken.symbol, 5)}
-      </span>
-    </div>
+    <TokenItem token={activeToken} />
   ) : (
-    <Tooltip
-      trigger="hover"
-      popperOptions={{ placement: 'bottom' }}
-      tooltipContent={formatText({ id: 'actionSidebar.tokenErrorTooltip' })}
-    >
-      <div className="flex items-center gap-1 text-negative-400">
-        <WarningCircle size={16} />
-        <span className="text-md">
-          {formatText({
-            id: 'actionSidebar.tokenError',
-          })}
-        </span>
-      </div>
-    </Tooltip>
+    <TokenNotFound />
   );
 
   return (
@@ -249,72 +222,12 @@ const AmountField: FC<AmountFieldProps> = ({
                     <h5 className="mb-2 ml-4 uppercase text-gray-400 text-4">
                       {formatText({ id: 'actionSidebar.availableTokens' })}
                     </h5>
-                    <ul>
-                      {colonyTokens.map((colonyToken) => {
-                        const tokenBalance = getBalanceForTokenAndDomain(
-                          colony.balances,
-                          colonyToken.tokenAddress,
-                          domainId ?? Id.RootDomain,
-                        );
-
-                        return (
-                          <li
-                            key={colonyToken.tokenAddress}
-                            className="mb-1 last:mb-0"
-                          >
-                            <HoverWidthWrapper hoverClassName="font-medium block">
-                              <button
-                                type="button"
-                                className={`flex w-full items-center justify-between
-                        gap-1 rounded-lg px-4 py-2 transition-colors
-                        md:hover:bg-gray-50 md:hover:font-medium`}
-                                onClick={() => {
-                                  tokenAddressController.onChange(
-                                    colonyToken.tokenAddress,
-                                  );
-                                  toggleTokenSelect();
-                                }}
-                              >
-                                <div className="flex items-center gap-1">
-                                  <TokenAvatar
-                                    size={18}
-                                    tokenName={colonyToken.name}
-                                    tokenAddress={colonyToken.tokenAddress}
-                                    tokenAvatarSrc={
-                                      colonyToken.avatar ?? undefined
-                                    }
-                                  />
-                                  <span className="text-md">
-                                    {multiLineTextEllipsis(
-                                      colonyToken.symbol,
-                                      5,
-                                    )}
-                                  </span>
-                                </div>
-                                {tokenBalance && (
-                                  <span className="ml-2 whitespace-nowrap text-sm text-gray-400">
-                                    {formatText({
-                                      id: 'actionSidebar.availableFunds',
-                                    })}
-                                    {': '}
-                                    <Numeral
-                                      value={tokenBalance}
-                                      decimals={getTokenDecimalsWithFallback(
-                                        colonyToken?.decimals,
-                                      )}
-                                    />{' '}
-                                    {multiLineTextEllipsis(
-                                      colonyToken.symbol,
-                                      5,
-                                    )}
-                                  </span>
-                                )}
-                              </button>
-                            </HoverWidthWrapper>
-                          </li>
-                        );
-                      })}
-                    </ul>
+                    <TokenList
+                      tokens={colonyTokens}
+                      domainId={domainId}
+                      balances={colony.balances}
+                      onSelect={handleTokenSelect}
+                    />
                   </MenuContainer>
                 </Portal>
               )}
