+++ src/components/v5/common/ActionSidebar/partials/AmountField/partials/TokenItem.tsx
@@ -1,3 +1,4 @@
+import clsx from 'clsx';
 import { type BigNumber } from 'ethers';
 import React, { type FC } from 'react';
 
@@ -10,12 +11,17 @@ import { TokenAvatar } from '~v5/shared/TokenAvatar/TokenAvatar.tsx';
 interface TokenItemProps {
   token: Token;
   tokenBalance?: BigNumber;
+  className?: string;
 }
 
-export const TokenItem: FC<TokenItemProps> = ({ token, tokenBalance }) => {
+export const TokenItem: FC<TokenItemProps> = ({
+  token,
+  tokenBalance,
+  className,
+}) => {
   return (
     <>
-      <div className="flex items-center gap-2">
+      <div className={clsx('flex items-center gap-1', className)}>
         <TokenAvatar
           size={18}
           tokenName={token.name}
